remove SyntaxError, SemanticError
