+++ src/index.ts
@@ -6,9 +6,8 @@ import { Interpreter } from './interpreter/index';
 import { Scope } from './interpreter/scope';
 import * as utils from './interpreter/util';
 import * as values from './interpreter/value';
-import { AiScriptError } from './interpreter/error';
 import { Parser, ParserPlugin, PluginType } from './parser';
-import { SyntaxError, SemanticError } from './error';
+import { AiScriptError } from './error';
 import * as N from './node';
 export { Interpreter };
 export { Scope };
@@ -18,6 +17,4 @@ export { AiScriptError };
 export { Parser };
 export { ParserPlugin };
 export { PluginType };
-export { SyntaxError };
-export { SemanticError };
 export { N };
